fixed errors and renamed
